Fixed example 4

+++ src/dtscalibration/plot.py
@@ -314,7 +314,7 @@ def plot_residuals_reference_sections_single(
     legend_ax = fig.add_subplot(grid[:2, :2], xticklabels=[], yticklabels=[])
     cbar_ax = fig.add_subplot(grid[2:, -1], xticklabels=[], yticklabels=[])
     if np.issubdtype(resid[time_dim].dtype, float) or np.issubdtype(
-            resid[time_dim].dtype, np.int):
+            resid[time_dim].dtype, int):
         resid.plot.imshow(
             ax=main_ax,
             cbar_ax=cbar_ax,
